[Mayank] Refactor Implemented ER Diagram and retrieved from multiple tables

+++ EmployeeModel.cs
@@ -8,12 +8,12 @@ namespace EmployeePayrollService
     {
         public int EmployeeID { get; set; }
         public string EmployeeName { get; set; }
-        public int Salary { get; set; }
+        public double Salary { get; set; }
         public string PhoneNumber { get; set; }
         public string Address { get; set; }
         public string Department { get; set; }
         public char Gender { get; set; }
-        public decimal BasicPay { get; set; }
+        public double BasicPay { get; set; }
         public double Deductions { get; set; }
         public double TaxablePay { get; set; }
         public double Tax { get; set; }
